djangoApiApp/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-environ==0.11.2
django-filter==23.2
django-core==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
install==1.3.5
Markdown==3.4.4
mysqlclient==2.2.0
packaging==23.1
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
